#117 Move battle tick lifecycle into battle::process_ticks
processing.rs now delegates to a single battle::process_ticks call.
All survivor-checking, conclude, and remove logic lives in tick.rs.
